Add factor fields to transactions
We update `ViewTransactionPresenter` to include `section130Factor` and `section127Factor` for both sroc and presroc transactions, and `winterOnlyFactor` for sroc only. As before, we need to extract factors for presroc from a string in the format `S127 x 0.5`. However for sroc, the factor values are extracted _before_ being persisted. We therefore extract the value from the string if the ruleset is presroc, or display the factor as-is for sroc.