Add factor values to View Invoice and Calculate Charge responses (#741)
* Add factor values to View Invoice and Calculate Charge responses

https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-310

We want to update the View Invoice and Calculate Charge responses to include factor values.

View Invoice for presroc is to include:
* section130Factor
* section127Factor

View Invoice for sroc is to include:
* winterOnlyFactor
* section130Factor
* section127Factor

Calculate Charge for presroc is to include:
* section130Factor
* section127Factor

Calculate Charge for sroc already includes the factors we are adding to View Invoice.

These are defined as follows:
* `section130Factor` for presroc is the number part of the rules service response entitled `licenceHolderChargeAgreement`. E.g. if the rules service returns`S130U x 0.5`, the `section130Factor` is `0.5`.
* `section127Factor` for presroc is the number part of the rules service response entitled `chargeElementAgreement`, where the first part of the string contains the text `S127`. E.g. if the rules service returns `S127 x 0.5`, the `section127Factor` is `0.5`. NB If the rules service response is `S126 x n.n`, then the `section127Factor` is `null`.
* `winterOnlyFactor`, `section130Factor` and `section127Factor` for sroc are the same as currently defined in the sroc Calculate Charge response.

Note that although it is the View Invoice response that we wish to change to include transaction factors, this presenter calls on the View Licence presenter which in turn calls on the View Transaction presenter. It is therefore `ViewTransactionPresenter` that we need to update to include the factors.

* Add `_extractFactor` method to `BasePresenter`

Because presroc data already exists in the db in the format `S127 x 0.5`, we're going to continue to store it in that format and do the conversion to a factor in the presenter. We start by creating an `_extractFactor` method to do this

* Add `_extractS127Factor` method to `BasePresenter`

Our presroc presenters will be receiving a string for `chargeElementAgreement` (which will have been persisted as `lineAttr10`) and only displaying the factor if this string starts with `S127`. We put the logic for this into `BasePresenter`

* Add fields to `CalculateChargePresrocPresenter`

We add the `section130Factor` and `section127Factor` fields to `CalculateChargePresrocPresenter`

* Extract factors with more than 1 dp

Based on what we'd seen in the ACs and in the water abstraction validation for incoming data, we made the assumption that a factor will have no more than 1 decimal place. However when we started investigating some failing tests, we realised that this assumption was incorrect, and that values such as `0.833` could potentially be returned by the Rules Service. We therefore update `_extractFactor()` to allow factors with more than 1 decimal place.

* Fix `CalculateChargeService` tests

Adding `section127Factor` and `section130Factor` to the calculate charge response broke some unit tests which did not expect these fields. We therefore update the fixtures to include them.

* Rename presenters to be more descriptive

* Add factor fields to transactions

We update `ViewTransactionPresenter` to include `section130Factor` and `section127Factor` for both sroc and presroc transactions, and `winterOnlyFactor` for sroc only. As before, we need to extract factors for presroc from a string in the format `S127 x 0.5`. However for sroc, the factor values are extracted _before_ being persisted. We therefore extract the value from the string if the ruleset is presroc, or display the factor as-is for sroc.

* Tidy up unit tests

We group the presroc and sroc tests together, and remove the test that checks the required fields are returned as this is implicitly tested by the "correctly presents the data" tests.