Change SRoC transaction file after user feedback (#743)
* Change SRoC transaction file after user feedback

https://eaflood.atlassian.net/jira/software/projects/CMEA/boards/907?selectedIssue=CMEA-309

The following changes have been requested for the sroc transaction file:

* Line attribute 1 (col 26) should be populated for all transactions -- a change from the current behaviour of populated for all except compensation charge transactions
* Line attribute 7 (col 32) and line attribute 11 (col 36) should all have `pence` appended without a space between eg. `969000pence`
* Line attribute 9 (col 34) should be in the format `${supportedSourceValue}pence (${supportedSourceName})` eg. `1881600pence (Kielder)`.
* Line attribute 10 (col 35) should be in the format `Calculated using reported abstracted quantity (${actualVolume}ML) and the charge reference, which is based on the authorised quantity`

* Update col 26 (`lineAttr1`)

Col 26 should now be populated for all transactions, not just compensation charge transactions. We therefore remove the conditionality.

* Add `pence` to col 32 and col 36

Cols 32 and 36 have `pence` appended to their output (without a space, eg. `12345pence`)

* Reformat col 34

Col 34 goes from being `value, supported_source_name` to `value (supported_source_name)`, with `pence` appended to the value

* Reformat col 35

Col 35 is changed to the string `Calculated using reported abstracted quantity (${actualVolume}ML) and the charge reference, which is based on the authorised quantity`

* Fix code smell

SonarCloud flagged up a code smell where a line was longer than 180 chars. This is due to col 35 now outpuitting a ~140 char string of text. To fix this, we split across a few lines the ternary that determines whether to use this string or a blank string.