Correctly return `section127Factor` for sroc invoices (#745)
https://eaflood.atlassian.net/browse/CMEA-310

We found in testing that `section127Factor` was not being correctly returned when an invoice was viewed. On investigation, we realised that we'd got our fields confused!

For presroc transactions, the field `lineAttr10` holds the factor string eg. `S127 x 0.5`. We mistakenly used this field when adding the factor to sroc transactions, when in fact we store the factor in `lineAttr15`. We therefore fix the transaction presenter to use the correct field.

While doing this, we realised that the two rulesets weren't consistent in how they return factors:
* For presroc we extract the factor from the string, parse it as a number and return it
* For sroc we return the value as-is, and since it's stored as a string (due to our use of generic fields in the db) this means we return it as a string.

We therefore make a further change to the transaction presenter to convert factors back from strings to floats.