Remove pushing to Docker Hub (#1059)
Access to push to Docker Hub was removed some time ago but we failed to realise that our `docker.yml` GitHub action was still trying to push to it. Because this step fails the rest of the workflow fails which means images are no longer being pushed to GHCR any more.

This change removes the login to Docker Hub step.

---

We also add a tweak to our `.gitignore` to keep out some files devs like to set up locally.