Fix incompatible npm version in Dockerfile
We've just got the GitHub action `.github/workflows/docker.yml` working again but it has highlighted a new issue. The version of Node the CHA is using is now a few versions behind. This means the latest versions of NPM are no longer compatible with it.

This change fixes the issue specifying the version of NPM to install.

> See https://nodejs.org/en/about/previous-releases for Node and NPM compatibility