Push docker image to AWS ECR
We're the last service using the Defra Artifactory and our web-ops friends would really like to retire it. [AWS ECR](https://aws.amazon.com/ecr/) is now the norm for services and they would really like it if we could migrate to using it instead.

The first step to that is confirming we can push our image to the registry they have configured.

This change updates our `docker.yml` GitHub action to push to ECR instead of GHCR, which hasn't been working for more than 2 months now.