Stop populating col35 in SROC 2PT transaction file (#1185)
https://eaflood.atlassian.net/browse/WATER-4740

This request has come from the [http://github.com/DEFRA/water-abstraction-team](Water Abstraction Team) to amend the transaction file generated for SROC bill runs.

When the transaction is flagged as being two-part tariff, we currently have logic that outputs an attempt to explain how the charge was calculated. However, if the charge involves aggregates, the explanation is wrong. Explaining the charge in a short sentence would be a project itself! So, they have requested we drop the output altogether and leave it blank.