Fix failing tests due to `mock-fs` dev dependency
Our tests fail wherever we use the `mock-fs` dev dependency. We update the depdency from 4.x.x to 5.x.x which resolves this.

However, some tests now fail because the `err.message` text `ENOENT: no such file or directory, open '${fakeFile}'` has changed to `ENOENT, no such file or directory '${fakeFile}'`.

We could simply change the test to look for the amended text; however to avoid this happening again, we change the test to indiviually check that `err.code` is `ENOENT` and `err.path` is our value of `fakeFile`.